[Solaredge] Clean up date normalization
The solaredge API has a special interpretation of the date ranges. The
first data point returned is the `start_time` rounded down to the full
quarter of an hour. The last data point returned is the `end_time` of
the last (not equal) full quarter of an hour. The last available data
point might change until the next quarter of the hour.

The later resulted in partially updated measurements and thus incorrect
statistics. Now the API returns always the last data point until it is
fully fixed.