[gardena] Run metric collection as daemon
In contrast to the other python scripts that collect metrics whenever
the script is run, the gardena script is waiting for the events from the
API. That script was started manually in a screen, but had to be
restarted whenever the raspberry was down. This commit adds a script
that can be scheduled to check and if necessary restart the script.